Include full post for domain modeling with sum types
We include the full body of the post for domain modeling with sum types
in elixir. But we add a comment at the top of the post to show where it
was originally published with a link to it. I think this provides a
better experience for the user so that they aren't having to jump back
and forth between thoughtbot's blog and mine if they want to read more
than one article.