Add full body to posts originally published in thoughtbot
I include the full body of the posts, but add a message at the top to
inform the reader where they can find the original if they want to. I
thin this provides a better experience for the user rather than being
given no option but to keep jumping to thoughtbot and then back in order
to read more than one post in my website.